requirements.txt
django 
djangorestframework
celery 
redis
psycopg2-binary
google-genai
python-dotenv
jinja2
daphne
pandas
openpyxl
